+++ src/algorithms/acquisition/adapters/glonass_l1_ca_pcps_acquisition.cc
@@ -55,7 +55,7 @@ GlonassL1CaPcpsAcquisition::GlonassL1CaPcpsAcquisition(
             acq_parameters_.doppler_max = FLAGS_doppler_max;
         }
     doppler_max_ = acq_parameters_.doppler_max;
-    doppler_step_ = acq_parameters_.doppler_step;
+    doppler_step_ = static_cast<unsigned int>(acq_parameters_.doppler_step);
     item_type_ = acq_parameters_.item_type;
     item_size_ = acq_parameters_.it_size;
     fs_in_ = acq_parameters_.fs_in;
