+++ src/algorithms/acquisition/adapters/gps_l1_ca_pcps_acquisition.cc
@@ -58,7 +58,7 @@ GpsL1CaPcpsAcquisition::GpsL1CaPcpsAcquisition(
         }
 
     doppler_max_ = acq_parameters_.doppler_max;
-    doppler_step_ = acq_parameters_.doppler_step;
+    doppler_step_ = static_cast<unsigned int>(acq_parameters_.doppler_step);
     item_type_ = acq_parameters_.item_type;
     item_size_ = acq_parameters_.it_size;
